refactor(telegram): extract watch command handler from bot.ts
bot.ts had grown to 394 code lines, exceeding the 392-line grandfathered
baseline (GREW violation). Extract the /watch command handler into its own
module to bring bot.ts back within the registered ceiling.

- New file: src/telegram/bot.watch-command.ts (57 lines)
  Contains the /watch command handler logic previously inlined in bot.ts.
- bot.ts: 394 → 369 code lines (25-line reduction)
- .filesize-baseline.json: update bot.ts entry 392 → 369
- .funcsize-baseline.json: updated to reflect new module boundaries

All gates pass: lint (tsc --noEmit), pnpm audit:filesize:check.
